moved tests and rocket to main
